client._ui: refactor line formatting
Refactor line formatting into a separate
class with a draw method to be called
from the ProgressBar class.

Format class is instantiated at ProgressBar
construction and allows to easily support
different format for the progress bar
to be printed. To that end, we only have
to inherit from PBFormat and override the
draw method.

Signed-off-by: Albert Esteve <aesteve@redhat.com>